requirements.txt
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
bidict==0.21.4
black==22.1.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.73.0
flake8==4.0.1
gevent==21.12.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.0
psycopg2==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyflakes==2.4.0
python-dotenv==0.19.2
python-engineio==4.3.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.1.0
uvicorn==0.17.6
watchgod==0.8
websockets==10.2
zope.event==4.5.0
zope.interface==5.4.0
